Extract app logic from index.js to app.js
